+++ CHANGELOG.md
@@ -51,6 +51,7 @@ title: Changelog
 - Introduced `@function` tag to force TypeDoc to convert variable declarations with a type annotation as functions, #2881.
 - Exposed a `TypeDoc` global object in the HTML theme which can be used to prevent TypeDoc from using `localStorage`, #2872.
 - Introduced `@preventInline` and `@inlineType` tags for further control extending the `@inline` tag, #2862.
+- API: Introduced `DefaultThemeRenderContext.reflectionIcon` for more granular control over displayed reflection icons.
 
 ### Bug Fixes